Update Catapult.java
Adding code to check the two switches

I have the feeling that the switches are wired reverse: i.e. it returns false when the limit switch is pressed and returns true when the limit switch is not pressed

Test by manually pressing the limit switch. and pressing winch button.  Make sure the println changes as you press/release the limit switch.

It's possible that they wired the thing into the side car and we have to change the module number.  (i.e. in the constructor, use module 2 instead of 1 in the "new DigitalInput" code